+++ binding/slice_binder.go
@@ -1,0 +1,106 @@
+package binding
+
+import (
+	"reflect"
+)
+
+var (
+	sliceBinder = Binder{
+		Bind:   bindSlice,
+		Unbind: unbindSlice,
+	}
+)
+
+
+// Used to keep track of the index for individual keyvalues.
+type sliceValue struct {
+	index int           // Index extracted from brackets.  If -1, no index was provided.
+	value reflect.Value // the bound value for this slice element.
+}
+
+
+// This function creates a slice of the given type, Binds each of the individual
+// elements, and then sets them to their appropriate location in the slice.
+// If elements are provided without an explicit index, they are added (in
+// unspecified order) to the end of the slice.
+func bindSlice(params *Params, name string, typ reflect.Type) reflect.Value {
+	// Collect an array of slice elements with their indexes (and the max index).
+	maxIndex := -1
+	numNoIndex := 0
+	sliceValues := []sliceValue{}
+
+	// Factor out the common slice logic (between form values and files).
+	processElement := func(key string, vals []string, files []*multipart.FileHeader) {
+		if !strings.HasPrefix(key, name+"[") {
+			return
+		}
+
+		// Extract the index, and the index where a sub-key starts. (e.g. field[0].subkey)
+		index := -1
+		leftBracket, rightBracket := len(name), strings.Index(key[len(name):], "]")+len(name)
+		if rightBracket > leftBracket+1 {
+			index, _ = strconv.Atoi(key[leftBracket+1 : rightBracket])
+		}
+		subKeyIndex := rightBracket + 1
+
+		// Handle the indexed case.
+		if index > -1 {
+			if index > maxIndex {
+				maxIndex = index
+			}
+			sliceValues = append(sliceValues, sliceValue{
+				index: index,
+				value: Bind(params, key[:subKeyIndex], typ.Elem()),
+			})
+			return
+		}
+
+		// It's an un-indexed element.  (e.g. element[])
+		numNoIndex += len(vals) + len(files)
+		for _, val := range vals {
+			// Unindexed values can only be direct-bound.
+			sliceValues = append(sliceValues, sliceValue{
+				index: -1,
+				value: BindValue(val, typ.Elem()),
+			})
+		}
+
+		for _, fileHeader := range files {
+			sliceValues = append(sliceValues, sliceValue{
+				index: -1,
+				value: BindFile(fileHeader, typ.Elem()),
+			})
+		}
+	}
+
+	for key, vals := range params.Values {
+		processElement(key, vals, nil)
+	}
+	for key, fileHeaders := range params.Files {
+		processElement(key, nil, fileHeaders)
+	}
+
+	resultArray := reflect.MakeSlice(typ, maxIndex+1, maxIndex+1+numNoIndex)
+	for _, sv := range sliceValues {
+		if sv.index != -1 {
+			resultArray.Index(sv.index).Set(sv.value)
+		} else {
+			resultArray = reflect.Append(resultArray, sv.value)
+		}
+	}
+
+	return resultArray
+}
+
+
+func unbindSlice(output map[string]string, name string, val interface{}) {
+	v := reflect.ValueOf(val)
+	for i := 0; i < v.Len(); i++ {
+		Unbind(output, fmt.Sprintf("%s[%d]", name, i), v.Index(i).Interface())
+	}
+}
+
+func init() {
+	KindBinders[reflect.Slice] = sliceBinder
+}
+
