+++ binding/struct_binder.go
@@ -1,0 +1,67 @@
+package binding
+
+import (
+	"reflect"
+)
+
+var (
+	structBinder = Binder{
+		Bind:   bindStruct,
+		Unbind: unbindStruct,
+	}
+)
+
+
+func bindStruct(params *Params, name string, typ reflect.Type) reflect.Value {
+	result := reflect.New(typ).Elem()
+	fieldValues := make(map[string]reflect.Value)
+	for key, _ := range params.Values {
+		if !strings.HasPrefix(key, name+".") {
+			continue
+		}
+
+		// Get the name of the struct property.
+		// Strip off the prefix. e.g. foo.bar.baz => bar.baz
+		suffix := key[len(name)+1:]
+		fieldName := nextKey(suffix)
+		fieldLen := len(fieldName)
+
+		if _, ok := fieldValues[fieldName]; !ok {
+			// Time to bind this field.  Get it and make sure we can set it.
+			fieldValue := result.FieldByName(fieldName)
+			if !fieldValue.IsValid() {
+				// TODO WARN.Println("W: bindStruct: Field not found:", fieldName)
+				continue
+			}
+			if !fieldValue.CanSet() {
+				// TODO WARN.Println("W: bindStruct: Field not settable:", fieldName)
+				continue
+			}
+			boundVal := Bind(params, key[:len(name)+1+fieldLen], fieldValue.Type())
+			fieldValue.Set(boundVal)
+			fieldValues[fieldName] = boundVal
+		}
+	}
+
+	return result
+}
+
+
+func unbindStruct(output map[string]string, name string, iface interface{}) {
+	val := reflect.ValueOf(iface)
+	typ := val.Type()
+	for i := 0; i < val.NumField(); i++ {
+		structField := typ.Field(i)
+		fieldValue := val.Field(i)
+
+		// PkgPath is specified to be empty exactly for exported fields.
+		if structField.PkgPath == "" {
+			Unbind(output, fmt.Sprintf("%s.%s", name, structField.Name), fieldValue.Interface())
+		}
+	}
+}
+
+func init() {
+	KindBinders[reflect.Struct] = structBinder
+}
+
